+++ src/TranslatorInterface.php
@@ -13,13 +13,17 @@ interface TranslatorInterface
 {
     /**
      * Add category sources.
+     *
+     * @return $this
      */
-    public function addCategorySources(CategorySource ...$categories): void;
+    public function addCategorySources(CategorySource ...$categories): static;
 
     /**
      * Set the default locale.
+     *
+     * @return $this
      */
-    public function setLocale(string $locale): void;
+    public function setLocale(string $locale): static;
 
     /**
      * @return string Default locale.
@@ -49,12 +53,14 @@ interface TranslatorInterface
     /**
      * Get a new translator instance with category to be used by default in case category isn't specified explicitly.
      *
-     * @psalm-immutable
+     * @return $this
      */
-    public function withDefaultCategory(string $category): self;
+    public function withDefaultCategory(string $category): static;
 
     /**
      * Get a new translator instance with locale to be used by default in case locale isn't specified explicitly.
+     *
+     * @return $this
      */
-    public function withLocale(string $locale): self;
+    public function withLocale(string $locale): static;
 }